Modularization of frontend components
